refactor proposal workflow routes
