+++ src/app/routers/proposals.py
@@ -1,8 +1,6 @@
 from fastapi import APIRouter, Header, Path, Query
 
 from app.contracts.proposals import (
-    ProposalApprovalActionRequest,
-    ProposalApprovalsEnvelopeResponse,
     ProposalBodyRequest,
     ProposalCreateEnvelopeResponse,
     ProposalCreateRequest,
@@ -10,7 +8,6 @@ from app.contracts.proposals import (
     ProposalDeliverySummaryEnvelopeResponse,
     ProposalDetailEnvelopeResponse,
     ProposalEnvelopeResponse,
-    ProposalLineageEnvelopeResponse,
     ProposalListEnvelopeResponse,
     ProposalNarrativeReviewEnvelopeResponse,
     ProposalNarrativeReviewRequest,
@@ -18,9 +15,6 @@ from app.contracts.proposals import (
     ProposalReportRequestEnvelopeResponse,
     ProposalSimulateRequest,
     ProposalSimulateResponse,
-    ProposalStateTransitionEnvelopeResponse,
-    ProposalSubmitRequest,
-    ProposalWorkflowEventsEnvelopeResponse,
 )
 from app.middleware.correlation import correlation_id_var
 from app.services.advisory_service_provider import proposal_service
@@ -192,198 +186,6 @@ async def get_proposal(
     return await service.get_proposal(
         proposal_id=proposal_id,
         include_evidence=include_evidence,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/submit",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Submit Proposal",
-    description="Submits a proposal into the next workflow review state.",
-)
-async def submit_proposal(
-    request: ProposalSubmitRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for proposal workflow submission requests.",
-        examples=["idem-submit-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.submit_proposal(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        review_type=request.review_type,
-        reason=request.reason,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/approve-risk",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Approve Proposal Risk Review",
-    description="Records the risk approval decision for a proposal in risk review.",
-)
-async def approve_risk(
-    request: ProposalApprovalActionRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for proposal risk-approval requests.",
-        examples=["idem-approve-risk-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.approve_risk(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        details=request.details,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/approve-compliance",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Approve Proposal Compliance Review",
-    description="Records the compliance approval decision for a proposal in compliance review.",
-)
-async def approve_compliance(
-    request: ProposalApprovalActionRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for proposal compliance-approval requests.",
-        examples=["idem-approve-compliance-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.approve_compliance(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        details=request.details,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.post(
-    "/{proposal_id}/record-client-consent",
-    response_model=ProposalStateTransitionEnvelopeResponse,
-    summary="Record Proposal Client Consent",
-    description="Records client consent for a proposal that has completed internal approvals.",
-)
-async def record_client_consent(
-    request: ProposalApprovalActionRequest,
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-    idempotency_key: str = Header(
-        alias="Idempotency-Key",
-        description="Caller-supplied idempotency key for client-consent recording requests.",
-        examples=["idem-client-consent-1"],
-    ),
-) -> ProposalStateTransitionEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.record_client_consent(
-        proposal_id=proposal_id,
-        actor_id=request.actor_id,
-        expected_state=request.expected_state,
-        details=request.details,
-        related_version_no=request.related_version_no,
-        idempotency_key=idempotency_key,
-        correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{proposal_id}/workflow-events",
-    response_model=ProposalWorkflowEventsEnvelopeResponse,
-    summary="Get Proposal Workflow Events",
-    description="Returns the workflow event timeline for a specific advisory proposal.",
-)
-async def get_workflow_events(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-) -> ProposalWorkflowEventsEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_workflow_events(
-        proposal_id=proposal_id,
-        correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{proposal_id}/approvals",
-    response_model=ProposalApprovalsEnvelopeResponse,
-    summary="Get Proposal Approvals",
-    description="Returns approval records already captured for a specific advisory proposal.",
-)
-async def get_approvals(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-) -> ProposalApprovalsEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_approvals(
-        proposal_id=proposal_id,
-        correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{proposal_id}/lineage",
-    response_model=ProposalLineageEnvelopeResponse,
-    summary="Get Proposal Lineage",
-    description="Returns immutable version lineage metadata and hashes for a specific proposal.",
-)
-async def get_proposal_lineage(
-    proposal_id: str = Path(
-        ...,
-        description="Gateway-visible proposal identifier returned by lotus-advise.",
-        examples=["pp_1"],
-    ),
-) -> ProposalLineageEnvelopeResponse:
-    service = proposal_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_proposal_lineage(
-        proposal_id=proposal_id,
         correlation_id=correlation_id,
     )
